fixed test review bug and renamed order route orders
