+++ server/app/routes/genres/index.js
@@ -3,10 +3,10 @@ var router = require('express').Router();
 module.exports = router;
 
 var mongoose = require('mongoose');
-var genre = mongoose.model('Genre')
+var Genre = mongoose.model('Genre')
 
 router.get('/', function(req, res) {
-	genre.find({}).exec()
+	Genre.find({}).exec()
 		.then(function(data) {
 			res.json(data);
 		})
@@ -16,7 +16,7 @@ router.get('/', function(req, res) {
 })
 
 router.get('/:genre', function(req, res) {
-	genre.find({
+	Genre.find({
 			name: req.params.genre
 		}).exec()
 		.then(function(data) {
