--- server/app/routes/order/index.js
@@ -1,7 +1,0 @@
-var router = require('express').Router();
-var order = require('./order.router');
-
-router.use('/', order);
-
-module.exports = router;
-
